requirements.txt
jpype1>=1.5.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
